// Qanat.Web/src/app/shared/generated/model/statement-batch-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class StatementBatchDto { 
    StatementBatchID?: number;
    StatementBatchName?: string | null;
    StatementTemplateName?: string | null;
    ReportingPeriodYear?: number;
    LastUpdated?: string;
    UpdateUserFullName?: string | null;
    NumberOfWaterAccounts?: number;
    StatementsGenerated?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface StatementBatchDtoForm { 
    StatementBatchID?: FormControl<number>;
    StatementBatchName?: FormControl<string>;
    StatementTemplateName?: FormControl<string>;
    ReportingPeriodYear?: FormControl<number>;
    LastUpdated?: FormControl<string>;
    UpdateUserFullName?: FormControl<string>;
    NumberOfWaterAccounts?: FormControl<number>;
    StatementsGenerated?: FormControl<boolean>;
}

export class StatementBatchDtoFormControls { 
    public static StatementBatchID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static StatementBatchName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static StatementTemplateName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ReportingPeriodYear = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LastUpdated = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static UpdateUserFullName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static NumberOfWaterAccounts = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static StatementsGenerated = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}